Added fallback for removed icon themes
For a few days, I had published a version with light/dark icon themes.
Then I published version 1.2, which replaced those themes with new ones,
as well as re-creating all the icons. This was a breaking change,
because people who updated would end up with an empty icon.

This commit fixes that issue by automatically switching back to the
default icon theme if an invalid icon theme was found.