v1.2.3: bugfix for the icon color not being set after a restart on Chrome
This is a mess. Why is it so hard to run something during the extension
initialization? The previous code was trivial and it worked flawlessly
on Firefox.

But on Chrome it didn't. Due to manifest v3 and the mandatory conversion
from background scripts to to service workers, the icon was being
changed once, but then during subsequent restarts it would revert back
to the default icon. That's because that code was executed only once.

Solution? Add the icon setup to one of the initialization events. But
which one? Well… That's the messy part. Read the comments in the code.
And, given the Firefox 136 still doesn't fully support service workers
in extensions, the old style of icon initialization cannot be removed.

It's messy.

It seems to be working, but I'm not sure I covered all the cases.